requirements.txt
langchain
langchain-community==0.3.1

pandas
tqdm

gensim
transformers
FlagEmbedding

redis
neo4j
faiss-gpu
yaml
pyyaml
numpy

metagpt

fastapi
